+++ odbc-api/src/guide.rs
@@ -60,7 +60,7 @@
 //!
 //!             // Use schema in cursor to initialize a text buffer large enough to hold the largest
 //!             // possible strings for each column.
-//!             let mut buffers = TextRowSet::new(BATCH_SIZE, &cursor)?;
+//!             let mut buffers = TextRowSet::for_cursor(BATCH_SIZE, &cursor)?;
 //!             // Bind the buffer to the cursor. It is now being filled with every call to fetch.
 //!             let mut row_set_cursor = cursor.bind_buffer(&mut buffers)?;
 //!